WIP on dyanic_base_class setup
