+++ toshi_hazard_store/model/__init__.py
@@ -7,13 +7,13 @@ from .gridded_hazard import GriddedHazard
 from .gridded_hazard import drop_tables as drop_gridded
 from .gridded_hazard import migrate as migrate_gridded
 from .location_indexed_model import LocationIndexedModel
-
-# from .openquake_models import tables as oqv3_tables
-# from .openquake_v2_model import
 from .openquake_models import VS30_KEYLEN, HazardAggregation, OpenquakeRealization, ToshiOpenquakeMeta
 from .openquake_models import drop_tables as drop_openquake
 from .openquake_models import migrate as migrate_openquake
 from .openquake_models import vs30_nloc001_gt_rlz_index
+
+# from .openquake_models import tables as oqv3_tables
+# from .openquake_v2_model import
 
 
 def migrate():
